// src/types/Lead.ts
export interface Lead {
  id: string;
  name: string;
  email: string;
  avatar: string;
  company?: string;
  title?: string;
  phone?: string;
  mobile?: string;
  website?: string;
  leadSource?: string;
  leadStatus?: string;
  industry?: string;
  annualRevenue?: string;
  createdBy?: string;
  skypeId?: string;
  modifiedBy?: string;
  createdTime?: string;
  modifiedTime?: string;
  twitter?: string;
  lastActivityTime?: string;
  leadOwner?: string;
}